+++ src/store.ts
@@ -142,7 +142,11 @@ export function useStore(
     }
   }
 
-  function setImportMap(map: ImportMap) {
+  function setImportMap(map: ImportMap, merge = false) {
+    if (merge) {
+      map = mergeImportMap(getImportMap(), map)
+    }
+
     if (map.imports)
       for (const [key, value] of Object.entries(map.imports)) {
         if (value) {
@@ -266,9 +270,15 @@ export function useStore(
       }
     }
     if (vueVersion.value) files._version = vueVersion.value
+    if (typescriptVersion.value !== 'latest' || files._tsVersion) {
+      files._tsVersion = typescriptVersion.value
+    }
     return '#' + utoa(JSON.stringify(files))
   }
-  const deserialize: ReplStore['deserialize'] = (serializedState: string) => {
+  const deserialize: ReplStore['deserialize'] = (
+    serializedState: string,
+    checkBuiltinImportMap = true,
+  ) => {
     if (serializedState.startsWith('#'))
       serializedState = serializedState.slice(1)
     let saved: any
@@ -282,9 +292,14 @@ export function useStore(
     for (const filename in saved) {
       if (filename === '_version') {
         vueVersion.value = saved[filename]
+      } else if (filename === '_tsVersion') {
+        typescriptVersion.value = saved[filename]
       } else {
         setFile(files.value, filename, saved[filename])
       }
+    }
+    if (checkBuiltinImportMap) {
+      applyBuiltinImportMap()
     }
   }
   const getFiles: ReplStore['getFiles'] = () => {
@@ -329,7 +344,7 @@ export function useStore(
   }
 
   if (serializedState) {
-    deserialize(serializedState)
+    deserialize(serializedState, false)
   } else {
     setDefaultFile()
   }
@@ -368,6 +383,7 @@ export function useStore(
     deleteFile,
     renameFile,
     getImportMap,
+    setImportMap,
     getTsConfig,
     serialize,
     deserialize,
@@ -436,9 +452,15 @@ export interface ReplStore extends UnwrapRef<StoreState> {
   deleteFile(filename: string): void
   renameFile(oldFilename: string, newFilename: string): void
   getImportMap(): ImportMap
+  setImportMap(map: ImportMap, merge?: boolean): void
   getTsConfig(): Record<string, any>
   serialize(): string
-  deserialize(serializedState: string): void
+  /**
+   * Deserializes the given string to restore the REPL store state.
+   * @param serializedState - The serialized state string.
+   * @param checkBuiltinImportMap - Whether to check the built-in import map. Default to true
+   */
+  deserialize(serializedState: string, checkBuiltinImportMap?: boolean): void
   getFiles(): Record<string, string>
   setFiles(newFiles: Record<string, string>, mainFile?: string): Promise<void>
 }